requirements.txt
# requirements.txt
gunicorn
Flask
Flask-Cors
werkzeug
spacy==3.7.2
sentence-transformers==2.2.2
keybert==0.7.0
PyMuPDF==1.23.8
python-docx==1.1.0
google-generativeai==0.3.2
scikit-learn==1.4.0
shap
matplotlib
numpy